toolkit/content/tests/widgets/test_moz_input_text.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>MozInputText Tests</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" href="chrome://global/skin/in-content/common.css">
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="module" src="chrome://global/content/elements/moz-input-text.mjs"></script>
  <script src="input-test-helpers.js"></script>
  <script class="testbody" type="application/javascript">
    let testHelpers = new InputTestHelpers();
    let html;

    add_setup(async function setup() {
      ({ html } = await testHelpers.setupLit());
      await testHelpers.setupInputTests({
        templateFn: (attrs, children) =>
          html`<moz-input-text ${attrs}>${children}</moz-input-text>`,
      });
    });

    add_task(async function testMozInputTextProperties() {
      await testHelpers.testCommonInputProperties("moz-input-text");
    });

    add_task(async function testMozInputTextEvents() {
      let { trackEvent, verifyEvents } = testHelpers.getInputEventHelpers();
      let target = await testHelpers.renderInputElements();
      let textInput = target.querySelector("moz-input-text");

      textInput.addEventListener("change", trackEvent);
      textInput.addEventListener("input", trackEvent);
      textInput.addEventListener("click", trackEvent);

      const TEST_STRING = "mozilla!";
      synthesizeMouseAtCenter(textInput.inputEl, {});
      sendString(TEST_STRING);
      textInput.blur();
      await TestUtils.waitForTick();

      // Verify that we emit input events for each char of the string,
      // and that change events are fired when the input loses focus.
      verifyEvents([
        { type: "click", localName: "moz-input-text", value: "" },
        ...Array.from(TEST_STRING).map((char, i) => ({
          type: "input",
          localName: "moz-input-text",
          value: TEST_STRING.substring(0, i + 1),
        })),
        { type: "change", localName: "moz-input-text", value: TEST_STRING },
      ]);
    });
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
